+++ cmd/gopact/main.go
@@ -24,7 +24,7 @@ const (
 	exitError = 1
 	exitUsage = 2
 
-	fallbackSDKVersion = "v0.0.20"
+	fallbackSDKVersion = "v0.0.21"
 	scaffoldGoVersion  = "1.25"
 )